\name{trainCens}
\docType{data}
\alias{trainCens}
\title{Sample Training Data for the Iterative BMA Algorithm for Survival Analysis}
\description{
This is an diffuse large B-cell lymphoma (DLBCL) dataset from Rosenwald et al. (2002).
This is a vector of censor information for 65 training samples. In this vector, 
0 = censored and 1 = uncensored.
}
\usage{data(trainCens)}
\format{The vector is called \code{trainCens}. 
}
\details{
} 
\source{
Full dataset: \url{http://llmpp.nih.gov/DLBCL/}.
}
\references{
Rosenwald, A., Wright, G., Wing, C., Connors, J., Campo, E. et al. (2002). 
The Use of Molecular Profiling to Predict Survival After Chemotherapy for Diffuse Large-B-Cell Lymphoma. 
The New England Journal of Medicine, 346(25), 1937-1947.
}
\keyword{datasets}

